# server/docs/swagger.yaml
definitions:
...
  orderManager.GoldShop:
    properties:
      ID:
        description: 主键ID
        type: integer
      buyId:
        description: 购买人Id
        type: integer
      createId:
        description: 创建人Id
        type: integer
      createTime:
        description: 创建时间
        type: string
      createdAt:
        description: 创建时间
        type: string
      goodsId:
        description: 商品Id
        type: integer
      updateId:
        description: 更新人Id
        type: integer
      updateTime:
        description: 更新时间
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.5.9
paths:
...
  /goldShop/createGoldShop:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建goldShop表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/orderManager.GoldShop'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建goldShop表
      tags:
      - GoldShop
  /goldShop/deleteGoldShop:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除goldShop表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/orderManager.GoldShop'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除goldShop表
      tags:
      - GoldShop
  /goldShop/deleteGoldShopByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除goldShop表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除goldShop表
      tags:
      - GoldShop
  /goldShop/findGoldShop:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 购买人Id
        in: query
        name: buyId
        type: integer
      - description: 创建人Id
        in: query
        name: createId
        type: integer
      - description: 创建时间
        in: query
        name: createTime
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 商品Id
        in: query
        name: goodsId
        type: integer
      - description: 更新人Id
        in: query
        name: updateId
        type: integer
      - description: 更新时间
        in: query
        name: updateTime
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询goldShop表
      tags:
      - GoldShop
  /goldShop/getGoldShopList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取goldShop表列表
      tags:
      - GoldShop
  /goldShop/updateGoldShop:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新goldShop表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/orderManager.GoldShop'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新goldShop表
      tags:
      - GoldShop
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"